chore: add `mainnet` suffix to solana and eclipse (#171)
### Description

This is a breaking change since `solana` and `eclipse` are renamed with
the `mainnet` suffix, to allow for a new core deployment. See
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/4410

### Backward compatibility

<!--
Are these changes backward compatible? Note that additions are backwards
compatible.

Yes/No
-->

### Testing

<!--
Have any new metadata configs and deployment addresses been used with
any Hyperlane tooling, such as the CLI?
-->

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>