fix: remove gasLimit overrides for rootstock/rootstocktestnet (#247)
### Description

fix: remove gasLimit overrides for rootstock/rootstocktestnet

hardcoding the gaslimit this high causes problems in the relayer, as it
will want the IGP payment to cover this limit. if the gas limit is set
to the block gas limit, this is likely never going to happen. so we
remove the overrides.

note, the overrides have been removed in the recently deployed agent
configs. this is to update the default registry metadata to match our
deployed agent setup.

### Backward compatibility

y

### Testing

manual