revert: "feat: Add gitleaks to husky precommit" (#582)
Reverts hyperlane-xyz/hyperlane-registry#578

because it's causing issues when importing in docs

https://yarnpkg.com/advanced/lifecycle-scripts#postinstall
> Postinstall scripts should be avoided at all cost, as they make
installs slower and riskier. Many users will refuse to install
dependencies that have postinstall scripts. Additionally, since the
output isn't shown out of the box, using them to print a message to the
user will not work as you expect.