chore: get non-fully connected routes in sync with actual deployment (#753)
### Description

This change sets `remoteRouters` to reflect non-fully connected
depoloyments in `-deploy.yaml` files. It also puts the `connections`
property in `-config.yaml` in sync.

This affects the following routes:

- `USDC/arbitrum-base-ethereum-lisk-optimism-polygon-zeronetwork`
- `USDT/arbitrum-ethereum-mantle-mode-polygon-scroll-zeronetwork`

### Drive-by Changes

- Updated the linter rules so that inline (flow-style) maps in YAML are
formatted correctly for both eslint and prettier.

### Related Issues

- https://abacus-network.pagerduty.com/incidents/Q1MSJBIKUABOPG
- ENG-1061

### Backward Compatibility

This change brings the registry in sync with actual deployments.

### Testing

Manually run `check-warp-deploy` on the routes in question.