fix: set correct reorgPeriod & blocktime (#904)
### Description
Blocktime of Starknet sepolia is 30s, not 5. Also the agents often see
errors of missing blocks, that is a result of the 0 reorgPeriod. Set
that to 1 to have some buffer. That missing buffer also results in the
scraper wrongfully not indexing TXs.
<!--
Summary of change.
Example: Add sepolia chain
-->

### Backward compatibility

<!--
Are these changes backward compatible? Note that additions are backwards
compatible.

Yes/No
-->

### Testing

<!--
Have any new metadata configs and deployment addresses been used with
any Hyperlane tooling, such as the CLI?
-->